fixed all reviews
