+++ steward/src/commands/schedule_corks/aave_cellar.rs
@@ -1,0 +1,9 @@
+mod shutdown;
+
+use abscissa_core::{clap::Parser, Command, Runnable};
+
+/// Schedule shutdown commands
+#[derive(Command, Debug, Parser, Runnable)]
+pub enum AaveCellarCmd {
+    Shutdown(shutdown::ShutdownCmd),
+}
